backend/requirements.txt
fastapi
uvicorn
python-multipart
pdfplumber
PyPDF2
httpx
python-jose[cryptography]
passlib[bcrypt]
python-dotenv
pydantic
pydantic-settings
